third_party/WebKit/LayoutTests/fast/canvas/canvas-hit-regions-basic-test-expected.txt

Hit detection and mouse event tests
PASS clickCanvas(100, 100) is 'face'
PASS clickCanvas(200, 200) is 'nose'
PASS clickCanvas(127, 242) is 'mouth'
PASS clickCanvas(150, 125) is 'eye'
PASS clickCanvas(250, 125) is 'eye'
PASS clickCanvas(200, 125) is 'face'
PASS clickCanvas(20, 10) is null

Hit detection and mouse/touch event tests with CSS sizing
PASS clickCanvas(100, 100) is 'face'
PASS clickCanvas(200, 200) is 'nose'
PASS clickCanvas(127, 242) is 'no event sent to canvas'
PASS clickCanvas(150, 125) is 'eye'
PASS clickCanvas(250, 125) is 'no event sent to canvas'
PASS clickCanvas(200, 125) is 'face'
PASS clickCanvas(20, 10) is null

Hit detection and mouse/touch event tests with adjusted pixel ratio
PASS clickCanvas(50, 50) is 'face'
PASS clickCanvas(100, 100) is 'nose'
PASS clickCanvas(64, 121) is 'mouth'
PASS clickCanvas(75, 63) is 'eye'
PASS clickCanvas(125, 63) is 'eye'
PASS clickCanvas(100, 63) is 'face'
PASS clickCanvas(10, 5) is null

Hit detection and mouse/touch event tests with adjusted page scale factor
PASS clickCanvas(50, 50, 0.5) is 'face'
PASS clickCanvas(100, 100, 0.5) is 'nose'
PASS clickCanvas(64, 121, 0.5) is 'mouth'
PASS clickCanvas(75, 63, 0.5) is 'eye'
PASS clickCanvas(125, 63, 0.5) is 'eye'
PASS clickCanvas(100, 63, 0.5) is 'face'
PASS clickCanvas(10, 5, 0.5) is null

NotSupportedError exception tests
PASS context.addHitRegion() threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ id : '' }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ id : undefined }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ control : {} }) threw exception TypeError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': member control is not of type Element..
PASS context.addHitRegion({ control : null }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ control : undefined }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ id : '', control : {} }) threw exception TypeError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': member control is not of type Element..
PASS context.addHitRegion({ id : undefined, control : {} }) threw exception TypeError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': member control is not of type Element..
PASS context.addHitRegion({ id : '', control : null }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ id : undefined, control : null }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ id : '', control : undefined }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..
PASS context.addHitRegion({ id : undefined, control : undefined }) threw exception NotSupportedError: Failed to execute 'addHitRegion' on 'CanvasRenderingContext2D': Both id and control are null..

PASS successfullyParsed is true

TEST COMPLETE
